Add CSS Style Sheet
This commit adds a CSS style sheet which is linked to each of the
pages on the site. The style sheet changes the a element to have a
background colour of yellow, with a text colour of blue.
It gives the h1 element a background colour of blue and a text colour of
white, and the h2 element a background colour of green with a text
colour of white. The README file has been updated to show these changes.